backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
python-dotenv==1.0.1
pydantic==2.10.3
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
PyJWT==2.9.0

# PostgreSQL support (optional - only install if using PostgreSQL)
# psycopg2-binary==2.9.9